refactor(NestedSelect): Move ItemRow in a specific component
